Add a config file and improve chat pulling
Don't rely on arkmanager's rcon tool to get chat messages. It requires
external scripts/cron etc. Instead use the RCONClient we made which is
significantly faster and allows us to decrease the asyncio sleep so we
are pulling chat more often.

The config file is pretty well documented, and is currently very much
subject to change.