More exception handling
Due to how we handle sending commands, it's possible that we disconnect
a socket prematurely. For example, if a server is rebooting, we could
attempt to send an authentication command and fail (closing the socket)
only to send the command we origionally wanted. This commit handles that
edge case since there is not a good way to detect socket-level
authentication failures... especially if the RCON socket is stale during
a reboot.